backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
watchdog>=4.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
